app/requirements.txt
# Pagrindinės bibliotekos
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
python-dotenv==1.0.1
werkzeug==3.0.1

# Duomenų apdorojimas
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.2
opencv-python==4.9.0.80

# Deep Learning
tensorflow==2.16.2
torch==2.2.1
torchvision==0.17.1
transformers==4.38.2

# Vizualizacija
matplotlib==3.8.3
seaborn==0.13.2

# Duomenų bazė
sqlalchemy==2.0.27

# Kiti įrankiai
pillow==10.2.0
tqdm==4.66.2 